호스팅으로 인한 파이어베이스, 파이어스토어, 스토리지 관련 생성 파일 gitignore 추가 및 회원 가입 페이지 접근 제한 해제

+++ src/components/mypage/PersonalPostingCard.jsx
@@ -15,10 +15,14 @@ const PersonalPostingCard = (item, url, text) => {
   const dispatch = useDispatch();
   const navigate = useNavigate();
   const deleteHandler = async () => {
-    await dispatch(deletePosting({posting_id: item.item.docID}));
-    dispatch(postingActions.setDefaultPostings());
-    dispatch(fsActions.setDefaultLastVisible());
-    dispatch(personalFetchPosting());
+    const result = window.confirm('정말 삭제하시겠습니까?');
+    if(result) {
+      await dispatch(deletePosting({posting_id: item.item.docID}));
+      dispatch(postingActions.setDefaultPostings());
+      dispatch(fsActions.setDefaultLastVisible());
+      dispatch(personalFetchPosting());
+    }
+    else return;
   }
 
     return (
